Move StatsSnapshot type to types package
Consolidate StatsSnapshot with other shared types in the types package for consistency and to prevent future circular import issues.
